Rename mock data file
